refactor: rename PlatformScrollView to ScrollView and move to its own file
